pyproject.toml
[project]
name = "concreto-rebelde"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=0.10.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "redis>=5.0.1",
]
